# openapi.yaml
openapi: 3.1.0

info:
  title: Todo API
  version: 1.0.0
  description: Personal task management API with user authentication and todo management.

servers:
  - url: https://api.todo.dangdoan.com
    description: Production
  - url: http://localhost:8080
    description: Local

security:
  - bearerAuth: []

tags:
  - name: Auth
    description: User registration, email verification, and login
  - name: Todos
    description: Create, read, update, and delete todos

paths:
  /auth/signup:
    post:
      summary: Sign up for a new account
      operationId: signup
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '202':
          description: Sign-up successful, verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
        '409':
          description: Email is already taken
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Conflict
                status: 409
                detail: Email 'user@example.com' is already taken.
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/verify:
    post:
      summary: Verify email with token
      operationId: verify
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired verification token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: about:blank
                title: Bad Request
                status: 400
                detail: Verification token is invalid or expired.
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/login:
    post:
      summary: Log in with email and password
      operationId: login
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or email not verified
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                invalidCredentials:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: Invalid email or password.
                emailNotVerified:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: Email not verified. Check your inbox for verification link.
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /todos:
    get:
      summary: List todos, optionally filtered by completion status
      operationId: listTodos
      tags: [Todos]
      parameters:
        - name: completed
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by completion status (true for completed, false for pending)
      responses:
        '200':
          description: Array of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new todo
      operationId: createTodo
      tags: [Todos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Todo identifier

    get:
      summary: Retrieve a specific todo by ID
      operationId: getTodo
      tags: [Todos]
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Update a todo's title or completion status
      operationId: updateTodo
      tags: [Todos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a specific todo by ID
      operationId: deleteTodo
      tags: [Todos]
      responses:
        '204':
          description: Todo deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    TodoResponse:
      type: object
      required: [id, title, completed, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
          examples: [1]
        title:
          type: string
          examples: ["Buy groceries"]
        completed:
          type: boolean
          examples: [false]
        created_at:
          type: string
          format: date-time
          examples: ["2026-04-28T10:00:00Z"]
        updated_at:
          type: string
          format: date-time
          examples: ["2026-04-28T10:00:00Z"]

    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          examples: ["Buy groceries"]

    UpdateTodoRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          examples: ["Buy groceries"]
        completed:
          type: boolean
          examples: [true]

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          examples: ["user@example.com"]
        password:
          type: string
          writeOnly: true
          examples: ["s3cr3t!!"]

    SignUpRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          examples: ["user@example.com"]
        password:
          type: string
          minLength: 8
          writeOnly: true
          examples: ["s3cr3t!!"]

    SignUpResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          examples: ["Verification email sent. Check your inbox to verify your account."]

    VerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          examples: ["550e8400-e29b-41d4-a716-446655440000"]

    AuthResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token validity in seconds
          examples: [3600]

    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
          description: >
            A URI reference identifying the problem type.
            Use "about:blank" for standard HTTP errors; omit if not extending the type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.

    ValidationError:
      type: object
      required: [detail, pointer]
      properties:
        detail:
          type: string
          description: Error message for this field
          examples: ["must be a positive integer"]
        pointer:
          type: string
          format: json-pointer
          description: JSON Pointer to the offending field
          examples: ["#/age"]

    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationError'

  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Unauthorized
            status: 401
            detail: Missing or invalid Bearer token.

    Forbidden:
      description: The authenticated user does not own this resource
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Forbidden
            status: 403
            detail: You do not have permission to access this todo.

    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Not Found
            status: 404
            detail: Todo with id 42 was not found.

    UnprocessableContent:
      description: Request body failed validation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
          example:
            type: https://api.todo.dangdoan.com/validation-error
            title: Your request is not valid.
            status: 422
            errors:
              - detail: must not be blank
                pointer: "#/title"
              - detail: size must be between 3 and 50
                pointer: "#/username"

    InternalServerError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred. Please try again later.